LayoutTests/plugins/bindings-test-expected.txt

ALERT: typeof(objCPlugin)=object
ALERT: objCPlugin()=TypeError
ALERT: typeof(objCPluginFunction)=function
ALERT: objCPluginFunction()=test
ALERT: typeof(cPlugin)=function
ALERT: cPlugin()=1
ALERT: typeof(cPlugin)=object
ALERT: cPlugin()=TypeError
This page tests invoking as functions JavaScript objects that are provided by plugins.
